% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{createEngine}
\alias{createEngine}
\title{Create a Zigzag-HMC engine object}
\usage{
createEngine(
  dimension,
  lowerBounds,
  upperBounds,
  seed,
  mean,
  precision,
  flags = 128L
)
}
\arguments{
\item{dimension}{the dimension of MTN.}

\item{lowerBounds}{a vector specifying the lower bounds.}

\item{upperBounds}{a vector specifying the upper bounds.}

\item{seed}{random seed.}

\item{mean}{the mean vector.}

\item{precision}{the precision matrix.}

\item{flags}{which SIMD instruction set to use. 128 = SSE, 256 = AVX.}
}
\value{
a list whose only element is the Zigzag-HMC engine object.
}
\description{
Create the C++ object to set up SIMD vectorization for speeding up calculations
for Zigzag-HMC ("Zigzag-HMC engine").
}
