% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hzzHelperFunctions.R
\name{.get_bound_time}
\alias{.get_bound_time}
\title{Title}
\usage{
.get_bound_time(position, velocity, constraits)
}
\arguments{
\item{position}{}

\item{velocity}{}
}
\value{

}
\description{
Title
}
