% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_initial_position}
\alias{.get_initial_position}
\title{Get an initial position for HZZ}
\usage{
.get_initial_position(p0, constraits)
}
\arguments{
\item{p0}{a d-dimensional vector of the initial value. It must satisfy all constraints. If not specified a random initial value will be used}

\item{constraits}{a (list?) of constraints}
}
\value{
a d-dimensional vector of the initial value
}
\description{
Get an initial position for HZZ
}
