% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hzzHelperFunctions.R
\name{.min_pos_root}
\alias{.min_pos_root}
\title{To find the minimal positive root of ax^2 + bx + c = 0 if it exists. Otherwise return Inf.}
\usage{
.min_pos_root(a, b, c)
}
\arguments{
\item{a}{vector of a}

\item{b}{vector of b}

\item{c}{vector of c}
}
\value{
vector of the minimal positive roots
}
\description{
To find the minimal positive root of ax^2 + bx + c = 0 if it exists. Otherwise return Inf.
}
