% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GenerateSample}
\alias{GenerateSample}
\title{Generate a sample from a truncated normal distribution.}
\usage{
GenerateSample(
  initial_position,
  initial_momentum,
  constraint_direc,
  constraint_row_normsq,
  constraint_bound,
  cholesky_factor,
  mean,
  total_time,
  prec_parametrized
)
}
\arguments{
\item{initial_position}{starting position}

\item{initial_momentum}{starting momentum}

\item{constraint_direc}{F matrix (k-by-d matrix where k is the number of
linear constraints)}

\item{constraint_row_normsq}{vector of squared row norms ofr constraint_direc}

\item{constraint_bound}{g vector (k dimensional)}

\item{cholesky_factor}{upper triangular matrix R from cholesky decomposition of
precision or covariance matrix into R^TR}

\item{mean}{mean of unconstrained Gaussian}

\item{total_time}{total time the particle will bounce for}

\item{prec_parametrized}{boolean for whether parametrization is by precision (true)
or covariance matrix (false)}
}
\value{
vector of position in standard normal frame
}
\description{
First "whiten" the constraints and starting position into the standard normal
frame, then generate a sample in that frame, and the convert back to the original
frame.
}
