% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonicHMC.R
\name{getHarmonicSample}
\alias{getHarmonicSample}
\title{One-step Harmonic HMC Sampler (Whitened Coordinates)}
\usage{
getHarmonicSample(
  whitenedPosition,
  whitenedConstraints,
  integrationTime,
  diagnosticMode = FALSE
)
}
\arguments{
\item{whitenedPosition}{Position in whitened coordinates}

\item{whitenedConstraints}{List from \code{applyWhitenTransform()}}

\item{integrationTime}{Time for dynamics simulation}

\item{diagnosticMode}{Return bounce diagnostics}
}
\description{
One-step Harmonic HMC Sampler (Whitened Coordinates)
}
