% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hzz.R
\name{hzz}
\alias{hzz}
\title{Title}
\usage{
hzz(get_prec_product, mean, position, momentum, t)
}
\arguments{
\item{get_prec_product}{function that returns the precision matrix - vector product}

\item{mean}{a d-dimensional mean vector}

\item{position}{a d-dimensional vector of the initial position}

\item{momentum}{a d-dimensional vector of the initial momentum}

\item{t}{time length to simulate the Markov process}
}
\value{

}
\description{
Title
}
