% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcmg.R
\name{rcmg}
\alias{rcmg}
\title{Sample from a constrained Gaussian distribution}
\usage{
rcmg(
  n,
  mean,
  cov = NULL,
  prec = NULL,
  lowerBounds,
  upperBounds,
  t,
  p0 = NULL,
  momentum = NULL,
  cppFlg = FALSE,
  nutsFlg = FALSE,
  rSeed = 666,
  randomFlg = TRUE,
  debugFlg = F
)
}
\arguments{
\item{n}{number of samples}

\item{mean}{a d-dimensional mean vector}

\item{prec}{a d-by-d precision matrix of the Gaussian distribution}

\item{t}{time length to simulate the Markov process}

\item{p0}{a d-dimensional vector of the initial value. It must satisfy all constraints. If not specified a random initial value will be used}

\item{cppFlg}{}

\item{constraits}{a d-dimensional vector where a positive (negative) number means >0 (<0) truncation.}
}
\value{
An n-by-d matrix where each row is a multivariate sample
}
\description{
Generate samples from a constrained Gaussian distribution with the Hamiltonian zigzag sampler.
}
