% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_hmc.R
\name{run_bouncy_sampler}
\alias{run_bouncy_sampler}
\title{Title}
\usage{
run_bouncy_sampler(
  n,
  initial_position,
  constraint_direc,
  constraint_bound,
  cholesky,
  mu,
  precision = TRUE,
  total_time = pi/2,
  seed = 1
)
}
\value{

}
\description{
Title
}
