% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonicHMC.R
\name{runHHMC}
\alias{runHHMC}
\title{Run Harmonic HMC Sampler}
\usage{
runHHMC(
  n,
  initialPosition,
  constraintDirec,
  constraintBound,
  choleskyFactor,
  unconstrainedMean,
  precParametrized = TRUE,
  integrationTime = c(pi/8, pi/2),
  seed = 1,
  diagnosticMode = FALSE
)
}
\arguments{
\item{n}{number of samples}

\item{initialPosition}{starting value for parameters}

\item{constraintDirec}{F matrix (k-by-d matrix where k is the number of
linear constraints)}

\item{constraintBound}{g vector (k dimensional)}

\item{choleskyFactor}{upper triangular matrix R from cholesky decomposition
of precision or covariance matrix into R^TR}

\item{unconstrainedMean}{mean of unconstrained Gaussian}

\item{precParametrized}{boolean for whether parametrization is by precision
(TRUE) or covariance matrix (FALSE)}

\item{integrationTime}{amount of time the particle travels for each sample.
Can either be a scalar value for a fixed time across all samples, or a length 2
vector of a lower and upper bound for uniform distribution from which the
bounce time is drawn from for each sample.}

\item{seed}{random seed}

\item{diagnosticMode}{boolean for whether to return the bounce distances for
each sample}
}
\value{
List of
"samples": d x n matrix of samples
"bounceDistances": list of bounces for each sample, only present if
diagnosticMode is TRUE
}
\description{
Sample from a truncated Gaussian distribution with constraints Fx+g >= 0.
}
\examples{
set.seed(1)
d = 10
A = matrix(runif(d^2)*2-1, ncol=d)
Sigma = t(A) \%*\% A
R = cholesky(Sigma)
mu = rep(0,d)
constraintDirec = diag(d)
constraintBound = rep(0,d)
initial = rep(1, d)
results = runHHMC(
100,
initial,
constraintDirec,
constraintBound,
R,
mu,
precParametrized = FALSE
)
}
