% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setMean}
\alias{setMean}
\title{Set the mean for the target MTN}
\usage{
setMean(engine, mean)
}
\arguments{
\item{engine}{A Zigzag-HMC engine container object.}

\item{mean}{the mean vector.}
}
\description{
Set the mean vector for a given Zigzag-HMC engine object.
}
