% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulateWhitenedDynamics}
\alias{simulateWhitenedDynamics}
\title{Simulate bouncing particle in whitened frame.}
\usage{
simulateWhitenedDynamics(
  initial_position,
  initial_momentum,
  constraint_direc,
  constraint_row_norm_sq,
  constraint_bound,
  total_time,
  diagnostic_mode
)
}
\arguments{
\item{initial_position}{starting position}

\item{initial_momentum}{starting momentum}

\item{constraint_direc}{F matrix (k-by-d matrix where k is the number of
linear constraints)}

\item{constraint_bound}{g vector (k dimensional)}

\item{total_time}{total time the particle will bounce for}

\item{constraint_row_normsq}{vector of squared row norms of constraint_direc}

\item{param}{diagnostic_mode boolean for whether to return the bounce
distances for each sample}
}
\value{
vector of position in standard normal frame
}
\description{
Simulate bouncing particle in whitened frame.
}
