% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{unwhitenPosition}
\alias{unwhitenPosition}
\title{Convert a position from standard normal frame back to original frame.}
\usage{
unwhitenPosition(
  position,
  cholesky_factor,
  unconstrained_mean,
  prec_parametrized
)
}
\arguments{
\item{position}{starting position}

\item{cholesky_factor}{upper triangular matrix R from cholesky decomposition
of precision or covariance matrix into R^TR}

\item{unconstrained_mean}{mean of unconstrained Gaussian}

\item{prec_parametrized}{boolean for whether parametrization is by
precision (true)
or covariance matrix (false)}
}
\value{
vector of position in original frame
}
\description{
Convert a position from standard normal frame back to original frame.
}
