% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{WhitenConstraints}
\alias{WhitenConstraints}
\title{Whiten constraints for use in GenerateUnwhitenedSample}
\usage{
WhitenConstraints(
  constraint_direc,
  constraint_bound,
  cholesky_factor,
  mean,
  prec_parametrized
)
}
\arguments{
\item{constraint_direc}{F matrix (k-by-d matrix where k is the number of
linear constraints)}

\item{constraint_bound}{g vector (k dimensional)}

\item{cholesky_factor}{upper triangular matrix R from cholesky decomposition of
precision or covariance matrix into R^TR}

\item{mean}{mean of unconstrained Gaussian}

\item{prec_parametrized}{boolean for whether parametrization is by precision (true)
or covariance matrix (false)}
}
\value{
List of new constraint directions, the squared row norms of those
constraints (for computational efficiency later), and new bounds
}
\description{
Transforms constraints of the form Fx+g >= 0 for a target normal distribution
into the corresponding constraints for a standard normal.
}
