% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{whitenPosition}
\alias{whitenPosition}
\title{Whiten a given position into the standard normal frame.}
\usage{
whitenPosition(
  position,
  constraint_direc,
  constraint_bound,
  cholesky_factor,
  unconstrained_mean,
  prec_parametrized
)
}
\arguments{
\item{position}{starting position}

\item{constraint_direc}{F matrix (k-by-d matrix where k is the number of
linear constraints)}

\item{constraint_bound}{g vector (k dimensional)}

\item{cholesky_factor}{upper triangular matrix R from cholesky decomposition
of precision or covariance matrix into R^TR}

\item{unconstrained_mean}{mean of unconstrained Gaussian}

\item{prec_parametrized}{boolean for whether parametrization is by
precision (true)
or covariance matrix (false)}

\item{momentum}{starting momentum}
}
\value{
vector of position in standard normal frame
}
\description{
Whiten a given position into the standard normal frame.
}
