% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{applyWhitenTransform}
\alias{applyWhitenTransform}
\title{Whiten constraints for use in generateUnwhitenedSample}
\usage{
applyWhitenTransform(
  constraintDirec,
  constraintBound,
  choleskyFactor,
  unconstrainedMean,
  precParametrized
)
}
\arguments{
\item{constraintDirec}{F matrix (k-by-d matrix where k is the number of
linear constraints)}

\item{constraintBound}{g vector (k dimensional)}

\item{choleskyFactor}{upper triangular matrix R from cholesky decomposition
of precision or covariance matrix into R^TR}

\item{unconstrainedMean}{mean of unconstrained Gaussian}

\item{precParametrized}{boolean for whether parametrization is by precision
(true) or covariance matrix (false)}
}
\value{
List of new constraint directions, the squared row norms of those
constraints (for computational efficiency later), and new bounds
}
\description{
Transforms constraints of the form Fx+g >= 0 for a target normal
distribution into the corresponding constraints for a standard normal.
}
