% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cholesky}
\alias{cholesky}
\title{Efficient Cholesky decomposition}
\usage{
cholesky(A)
}
\arguments{
\item{A}{matrix to decompose}
}
\value{
upper triangular matrix R such that A = U'U.
}
\description{
Compute Cholesky decomposition of a matrix.
}
