% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cholesky}
\alias{cholesky}
\title{Compute Cholesky decomposition of a matrix.}
\usage{
cholesky(A)
}
\arguments{
\item{A}{matrix to decompose}
}
\value{
upper triangular matrix R such that A = R'R.
}
\description{
Compute Cholesky decomposition of a matrix.
}
