% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{createEngine}
\alias{createEngine}
\title{Create ZigZag engine object}
\usage{
createEngine(dimension, lowerBounds, upperBounds, flags, info, seed)
}
\arguments{
\item{dimension}{the dimension of MTN (d)}

\item{lowerBounds}{the d-dimensional lower bound}

\item{upperBounds}{the d-dimensional upper bound}

\item{flags}{todo}

\item{info}{todo}

\item{seed}{random seed}
}
\value{
a zigzag engine object.
}
\description{
Create ZigZag engine object
}
