% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{createEngine}
\alias{createEngine}
\title{Create the C++ object}
\usage{
createEngine(dimension, lowerBounds, upperBounds, flags, seed, mean, precision)
}
\arguments{
\item{flags}{128}

\item{seed}{random seed}

\item{the}{dimension of MTN (d)}

\item{a}{d-dimensional vector specifying the upper bounds.}

\item{info}{1}
}
\value{
a zigzag engine object.
}
\description{
Create the C++ object
}
