% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{createEngine}
\alias{createEngine}
\title{Create ZigZag engine object}
\usage{
createEngine(dimension, mask, lowerBounds, upperBounds, flags, info, seed)
}
\arguments{
\item{locationCount}{Number of locations and size of distance matrix.}

\item{tbb}{Number of CPU cores to be used.}

\item{simd}{For CPU implementation: no SIMD (\code{0}), SSE (\code{1}) or AVX (\code{2}).}

\item{truncation}{Likelihood includes truncation term? Defaults to \code{TRUE}.}

\item{gpu}{Which GPU to use? If only 1 available, use \code{gpu=1}. Defaults to \code{0}, no GPU.}

\item{single}{Set \code{single=1} if your GPU does not accommodate doubles.}
}
\value{
zigZag engine object.
}
\description{
Helper function creates zigZag engine object with given latent dimension, location count and various
implementation details. Called by \code{MassivezigZag::engineInitial()}.
}
