% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{createNutsEngine}
\alias{createNutsEngine}
\title{Create ZigZag nuts engine object}
\usage{
createNutsEngine(
  dimension,
  lowerBounds,
  upperBounds,
  flags,
  info,
  seed,
  randomFlg,
  stepSize,
  mean,
  precision
)
}
\arguments{
\item{dimension}{the dimension of MTN (d)}

\item{lowerBounds}{the d-dimensional lower bound}

\item{upperBounds}{the d-dimensional upper bound}

\item{flags}{todo}

\item{info}{todo}

\item{seed}{random seed}

\item{randomFlg}{todo}

\item{stepSize}{todo}

\item{mean}{todo}

\item{precision}{todo}
}
\value{
a zigzag-nuts engine object.
}
\description{
Helper function creates zigZag nuts engine object with given latent dimension, location count and various
implementation details.
}
