% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSampleFunctions.R
\name{drawLaplaceMomentum}
\alias{drawLaplaceMomentum}
\title{Draw a random Laplace momentum}
\usage{
drawLaplaceMomentum(dim)
}
\arguments{
\item{dim}{dimension of the momentum.}
}
\value{
a d-dimensional Laplace-distributed momentum.
}
\description{
Draw a random Laplace momentum
}
