% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{drawLaplaceMomentum}
\alias{drawLaplaceMomentum}
\title{Draw a random Laplace momentum}
\usage{
drawLaplaceMomentum(d)
}
\arguments{
\item{d}{dimension of the momentum.}
}
\value{
a d-dimensional Laplace-distributed momentum.
}
\description{
Generate a d-dimensional momentum where the density of each element is proportional to exp(-|pi|).
}
\examples{
# Draw a 3-dimensional Laplace momentum with reproducible results
set.seed(3)
momentum <- drawLaplaceMomentum(3)
momentum
}
\seealso{
\code{\link[=zigzagHMC]{zigzagHMC()}}
}
