% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zigzagHMC.R
\name{getInitialPosition}
\alias{getInitialPosition}
\title{A function to get an eligible initial value for a MTN}
\usage{
getInitialPosition(mean, lowerBounds, upperBounds)
}
\arguments{
\item{mean}{a d-dimensional mean vector}

\item{lowerBounds}{a d-dimensional lower bound}

\item{upperBounds}{a d-dimensional lower bound}
}
\value{
an eligible d-dimensional initial value
}
\description{
A function to get an eligible initial value for a MTN
}
