% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSampleFunctions.R
\name{getSample}
\alias{getSample}
\title{Title}
\usage{
getSample(position, momentum, t, nutsFlg, engine)
}
\arguments{
\item{position}{a d-dimensional vector of the initial position}

\item{momentum}{a d-dimensional vector of the initial momentum}

\item{t}{time length to simulate the Markov process}

\item{nutsFlg}{todo}

\item{engine}{todo}
}
\value{
one MCMC sample from the target MTN
}
\description{
Title
}
