% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSampleFunctions.R
\name{getZigzagSample}
\alias{getZigzagSample}
\title{Generate one random sample from the target MTN with Zigzag-HMC or Zigzag-NUTS}
\usage{
getZigzagSample(position, momentum, nutsFlg, engine, stepZZHMC = NULL)
}
\arguments{
\item{position}{a d-dimensional initial position vector.}

\item{momentum}{a d-dimensional initial momentum vector.}

\item{nutsFlg}{logical. If \code{TRUE} the No-U-Turn sampler will be used (Zigzag-NUTS).}

\item{engine}{list. Its \code{engine} element is a pointer to the C++ object that
implements fast computations for Zigzag-HMC and Zigzag-NUTS.
or Zigzag-NUTS computing engine.}

\item{stepZZHMC}{step size for Zigzag-HMC. If \code{nutsFlg = TRUE}, \code{engine} contains
the base step size for Zigzag-NUTS).}
}
\value{
one MCMC sample from the target MTN
}
\description{
Generate one random sample from the target MTN with Zigzag-HMC or Zigzag-NUTS
}
