% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcmg.R
\name{rcmg}
\alias{rcmg}
\title{Sample from a constrained Gaussian distribution}
\usage{
rcmg(
  n,
  mean,
  cov,
  prec = NULL,
  lowerBounds,
  upperBounds,
  x = NULL,
  forcedStep = NULL,
  momentum = NULL,
  nutsFlg = FALSE,
  rSeed = 666,
  randomFlg = TRUE
)
}
\arguments{
\item{n}{number of samples}

\item{mean}{a d-dimensional mean vector}

\item{cov}{todo}

\item{prec}{a d-by-d precision matrix of the Gaussian distribution}

\item{lowerBounds}{todo}

\item{upperBounds}{todo}

\item{x}{a d-dimensional vector of the initial value. It must satisfy all constraints. If not specified a random initial value will be used}

\item{forcedStep}{todo}

\item{momentum}{todo}

\item{nutsFlg}{todo}

\item{rSeed}{todo}

\item{randomFlg}{todo}
}
\value{
An n-by-d matrix where each row is a multivariate sample
}
\description{
Generate samples from a constrained Gaussian distribution with the Hamiltonian zigzag sampler.
}
