% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcmg.R
\name{rcmg}
\alias{rcmg}
\title{Sample from a constrained Gaussian distribution}
\usage{
rcmg(n, mean, prec, constraits, t, burnin, p0 = NULL)
}
\arguments{
\item{n}{number of samples}

\item{mean}{a d-dimensional mean vector}

\item{prec}{a d-by-d precision matrix of the Gaussian distribution}

\item{constraits}{a (list?) of constraints}

\item{t}{time length to simulate the Markov process}

\item{burnin}{the number of burn-in iterations}

\item{p0}{a d-dimensional vector of the initial value. It must satisfy all constraints. If not specified a random initial value will be used}
}
\value{
An n-by-d matrix where each row is a multivariate sample
}
\description{
Generate samples from a constrained Gaussian distribution with the Hamiltonian zigzag sampler.
}
\examples{
rcmg(1,1,1,1,1)
}
