% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setMean}
\alias{setMean}
\title{Set mean for MTN}
\usage{
setMean(sexp, mean)
}
\arguments{
\item{sexp}{pointer to zigzag object}

\item{mean}{a numeric vector containing the MTN mean}
}
\description{
Set mean for MTN
}
