% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setPrecision}
\alias{setPrecision}
\title{Set the precision matrix for the target MTN}
\usage{
setPrecision(engine, precision)
}
\arguments{
\item{engine}{A Zigzag-HMC engine container object.}

\item{precision}{the precision matrix.}
}
\description{
Set the precision matrix for a given Zigzag-HMC engine object.
}
