% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setPrecision}
\alias{setPrecision}
\title{Set the precision matrix for the target MTN}
\usage{
setPrecision(sexp, precision)
}
\arguments{
\item{sexp}{pointer to zigzag object}

\item{precision}{the MTN precision matrix}
}
\description{
Set the precision matrix for the target MTN
}
