% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulateWhitenedDynamics}
\alias{simulateWhitenedDynamics}
\title{Simulate bouncing particle in whitened frame.}
\usage{
simulateWhitenedDynamics(
  initialPosition,
  initialMomentum,
  constraintDirec,
  constraintRowNormSq,
  constraintBound,
  integrationTime,
  diagnosticMode
)
}
\arguments{
\item{initialPosition}{starting position}

\item{initialMomentum}{starting momentum}

\item{constraintDirec}{F matrix (k-by-d matrix where k is the number of
linear constraints)}

\item{constraintRowNormSq}{vector of squared row norms of constraintDirec}

\item{constraintBound}{g vector (k dimensional)}

\item{integrationTime}{total time the particle will bounce for}

\item{param}{diagnosticMode boolean for whether to return the bounce
distances for each sample}
}
\value{
vector of position in standard normal frame
}
\description{
Simulate bouncing particle in whitened frame.
}
