% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{whitenPosition}
\alias{whitenPosition}
\title{Whiten a given position into the standard normal frame.}
\usage{
whitenPosition(
  position,
  constraintDirec,
  constraintBound,
  choleskyFactor,
  unconstrainedMean,
  precParametrized
)
}
\arguments{
\item{position}{starting position}

\item{constraintDirec}{F matrix (k-by-d matrix where k is the number of
linear constraints)}

\item{constraintBound}{g vector (k dimensional)}

\item{choleskyFactor}{upper triangular matrix R from cholesky decomposition
of precision or covariance matrix into R^TR}

\item{unconstrainedMean}{mean of unconstrained Gaussian}

\item{precParametrized}{boolean for whether parametrization is by
precision (true)
or covariance matrix (false)}
}
\value{
vector of position in standard normal frame
}
\description{
Whiten a given position into the standard normal frame.
}
